Arranged the menu items.
INexecService.cancel() is same as INexecService.quit(), so I replaced
the item to cancel with an item to quit.

This patch also changes the menu to enable/disable items.

This patch includes arranging how to store/know the session id. Because
the previous way was not good for the above changing.